backend/authentication-service/requirements.txt
Flask
Flask-Cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
flasgger
PyJWT
Werkzeug
pymongo
